mythmetadatalookup: rename commandlineparser.* to mythmetadatalookup_…
